#76 [add] BasicDialog
